refactor: split MainWindow.xaml.cs — extract backdrop/theme to partial class
Moved all backdrop, theme, and background image logic (~250 lines) from
MainWindow.xaml.cs to MainWindow.Backdrop.cs (partial class).
MainWindow.xaml.cs now only contains navigation and window lifecycle.
